+++ src/lsp/superbol_free_lib/vscode_extension.ml
@@ -308,13 +308,19 @@ let contributes =
               is to use dash-separated words in lowercase in order to keep
               consistency with GnuCOBOL's configurations.  *)
 
-           Manifest.PROPERTY.enum "superbol.cobol.dialect"
-             ~cases:Cobol_config.DIALECT.all_canonical_names
+           Manifest.PROPERTY.string "superbol.cobol.dialect"
              ~markdownDescription:
                (with_superbol_toml_note
-                  "Default COBOL dialect; \"default\" is equivalent to \
-                   \"gnucobol\".")
-             ~default:(`String Cobol_config.(DIALECT.to_string Default))
+                  @@ Pretty.to_string "Default COBOL dialect; `default` is equivalent to `gnucobol`. \
+                  The available dialects are %a. \
+                  The `-strict` version of a dialect permits using reserved words from \
+                  all other dialects as custom COBOL words, as the non-strict version also \
+                  reserve these words. \
+                  You may also specify an absolute path to a custom GnuCOBOL dialect \
+                  configuration file (see `cobc`'s `-conf` option)."
+                   Fmt.(list ~sep:(any ",@ ") (quote ~mark:"`" string))
+                   Cobol_config.DIALECT.all_canonical_names)
+             ~default:"default"
              ~order:1;
 
            Manifest.PROPERTY.enum "superbol.cobol.sourceFormat"
@@ -596,7 +602,7 @@ let contributes =
       Manifest.tomlValidation
         ~fileMatch:"superbol.toml"
         (* TODO: change this address to a more permanent one; also, substitute `master` for a version tag *)
-        ~url:"https://raw.githubusercontent.com/OCamlPro/superbol-studio-oss/master/schemas/superbol-schema-0.1.4.json";
+        ~url:"https://raw.githubusercontent.com/OCamlPro/superbol-studio-oss/master/schemas/superbol-schema-0.2.3.json";
     ]
     ~menus: [
       "editor/context",